+++ R/plot_response_curves.R
@@ -79,6 +79,9 @@ plot_response_curves <- function(
   #getting the training data
   data <- model$ranger.arguments$data
 
+  # Drop geometry before processing
+  data <- drop_geometry_if_sf(data)
+
   #getting the response variable
   response.variable <- model$ranger.arguments$dependent.variable.name